+++ projects/ontimize-web-ngx-report/src/lib/services/index.ts
@@ -1,4 +1,4 @@
 export * from './o-alert.service';
 export * from './o-report.service';
-export * from './reports-on-demand.service';
-export * from './reports.service';
+export * from './ontimize-report.service';
+export * from './o-report.service';
