+++ projects/ontimize-web-ngx-report/src/lib/services/o-report.service.ts
@@ -1,13 +1,117 @@
-import { HttpEventType, HttpHeaders, HttpRequest } from "@angular/common/http";
 import { Injectable, Injector } from "@angular/core";
-import { IFileService, Observable, OntimizeEEService, Util } from 'ontimize-web-ngx';
+import { OReportParam } from "../types/report-param.type";
+import { Observable, OErrorDialogManager, OntimizeEEService, ServiceRequestParam, ServiceResponse, Util } from 'ontimize-web-ngx';
+import { HttpEventType, HttpHeaders, HttpRequest } from "@angular/common/http";
 import { share } from 'rxjs/operators';
 
-@Injectable()
-export class OReportService extends OntimizeEEService implements IFileService {
-
+@Injectable({ providedIn: 'root' })
+export class OReportService extends OntimizeEEService {
+  public oErrorDialogManager: OErrorDialogManager;
   constructor(protected injector: Injector) {
     super(injector);
+    super.configureService(this.getDefaultServiceConfiguration());
+    this.oErrorDialogManager = injector.get(OErrorDialogManager);
+  }
+
+  public createReport(reportparams: OReportParam): Observable<any> {
+
+    const body = JSON.stringify(
+      reportparams
+    )
+    const url = this.urlBase + '/dynamicjasper/report';
+
+    return this.doRequest({
+      method: 'POST',
+      url: url,
+      body: body
+    });
+  }
+
+  public saveAsPreferences(preferencesparams?: object): Observable<any> {
+    const body = JSON.stringify(
+      preferencesparams
+    )
+    const url = this.urlBase + '/preferences/save';
+
+    return this.doRequest({
+      method: 'POST',
+      url: url,
+      body: body
+    });
+  }
+
+  public savePreferences(id: number, preferencesparams?: object): Observable<any> {
+    const body = JSON.stringify(
+      preferencesparams
+    )
+    const url = this.urlBase + '/preferences/update/' + id;
+
+    return this.doRequest({
+      method: 'PUT',
+      url: url,
+      body: body
+    });
+  }
+
+  public getPreferences(entity?: string, service?: string): Observable<any> {
+
+    const url = this.urlBase + '/preferences/preferences?entity=' + entity + '&service=' + service;
+
+    return this.doRequest({
+      method: 'GET',
+      url: url
+    });
+
+  }
+  public getFunctions(functionparams?: object): Observable<any> {
+
+    const body = JSON.stringify(
+      functionparams
+    )
+    const url = this.urlBase + '/dynamicjasper/functionsName';
+
+    return this.doRequest({
+      method: 'POST',
+      url: url,
+      body: body
+    });
+  }
+
+  public deletePreferences(id?: number): Observable<any> {
+
+    const url = this.urlBase + '/preferences/remove/' + id;
+    return this.doRequest({
+      method: 'DELETE',
+      url: url
+    });
+
+  }
+
+  /** overridden method to add error callback for all requests */
+  doRequest(param: ServiceRequestParam): Observable<ServiceResponse> {
+    return super.doRequest({
+      method: param.method,
+      url: param.url,
+      body: param.body,
+      errorCallBack: this.errorCallBack
+    });
+  }
+
+
+  errorCallBack(httpErrorResponse: any) {
+    const error = httpErrorResponse.error;
+    if (Util.isObject(error)) {
+      if (error['code'] === 1 && Util.isDefined(error['message'])) {
+        this.showNotificationError(error['message']);
+        return;
+      }
+    }
+    this.showNotificationError('MESSAGES.ERROR_QUERY');
+
+  }
+
+  showNotificationError(error: string) {
+    this.oErrorDialogManager.openErrorDialog(error);
   }
 
   upload(files: any[], entity: string, data?: object): Observable<any> {
@@ -73,14 +177,7 @@ export class OReportService extends OntimizeEEService implements IFileService {
       observer.error('Service unavailable');
     }
   }
-  protected buildHeadersReport(): HttpHeaders {
-    let headers = new HttpHeaders({ 'Access-Control-Allow-Origin': '*' });
-    const sessionId = this.authService.getSessionInfo().id;
-    if (Util.isDefined(sessionId)) {
-      headers = headers.append('Authorization', 'Bearer ' + sessionId);
-    }
-    return headers;
-  }
+
 
   public advancedQuery(_kv?: Object, _av?: Array<string>, entity?: string, _sqltypes?: Object, offset?: number, _pagesize?: number, _orderby?: Array<Object>): Observable<any> {
     offset = (Util.isDefined(offset)) ? offset : this.offset;
@@ -155,5 +252,13 @@ export class OReportService extends OntimizeEEService implements IFileService {
     });
   }
 
+  protected buildHeadersReport(): HttpHeaders {
+    let headers = new HttpHeaders({ 'Access-Control-Allow-Origin': '*' });
+    const sessionId = this.authService.getSessionInfo().id;
+    if (Util.isDefined(sessionId)) {
+      headers = headers.append('Authorization', 'Bearer ' + sessionId);
+    }
+    return headers;
+  }
 
 }
